-abstracted controls and transition slides from main.py to their own files, controls.py and transition.py respectively -general code cleanup in main.py
